requirements.txt
Flask
transformers
spacy
torch
elasticsearch
numpy
textblob
scrapy
fastapi
aiohttp
aiosqlite
python-dotenv
discord.py
pygame
pyttsx3
pyautogui
pillow
pynput
imageio
pydantic
